refactor(wallets): extract the Rango-to-CAIP chain mapping from wallets-shared
Move the chain-vocabulary translation into `@rango-dev/wallets-blockchains` and
the CAIP-2 registry it uses into `@rango-dev/internal-blockchains`, so provider metadata and
namespace code depend only on the CAIP half and signers on the Rango half.
`Networks` moves with the mapping and wallets-shared re-exports it;
