+++ blockchains/internal-blockchains/src/networks.ts
@@ -1,0 +1,47 @@
+export enum Networks {
+  BTC = 'BTC',
+  BSC = 'BSC',
+  LTC = 'LTC',
+  THORCHAIN = 'THOR',
+  BCH = 'BCH',
+  BINANCE = 'BNB',
+  BLAST = 'BLAST',
+  CELO = 'CELO',
+  LINEA = 'LINEA',
+  METIS = 'METIS',
+  SONIC = 'SONIC',
+  BERACHAIN = 'BERACHAIN',
+  ZETA_CHAIN = 'ZETA_CHAIN',
+  SCROLL = 'SCROLL',
+  ZKSYNC = 'ZKSYNC',
+  MONAD = 'MONAD',
+  ETHEREUM = 'ETH',
+  POLYGON = 'POLYGON',
+  POLKADOT = '',
+  TRON = 'TRON',
+  DOGE = 'DOGE',
+  HARMONY = 'HARMONY',
+  AVAX_CCHAIN = 'AVAX_CCHAIN',
+  FANTOM = 'FANTOM',
+  MOONBEAM = 'MOONBEAM',
+  ARBITRUM = 'ARBITRUM',
+  BOBA = 'BOBA',
+  OPTIMISM = 'OPTIMISM',
+  FUSE = 'FUSE',
+  CRONOS = 'CRONOS',
+  SOLANA = 'SOLANA',
+  MOONRIVER = 'MOONRIVER',
+  GNOSIS = 'GNOSIS',
+  MAYA = 'MAYA',
+  STARKNET = 'STARKNET',
+  STELLAR = 'STELLAR',
+  TON = 'TON',
+  BASE = 'BASE',
+  SUI = 'SUI',
+  XRPL = 'XRPL',
+  ZCASH = 'ZCASH',
+  HYPERLIQUID = 'HYPERLIQUID',
+
+  // Using instead of null
+  Unknown = 'Unkown',
+}
