+++ wallets/provider-bitget/package.json
@@ -27,6 +27,7 @@
     "@hub3js/std": "^0.3.0",
     "@rango-dev/signer-evm": "^0.45.0",
     "@rango-dev/signer-tron": "^0.42.0",
+    "@rango-dev/internal-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.1-next.2",
     "rango-types": "^0.5.0"
