+++ wallets/provider-noir-wallet/package.json
@@ -24,6 +24,7 @@
     "@hub3js/core": "^0.61.1",
     "@hub3js/std": "^0.3.0",
     "@noir-wallet/sdk": "^0.1.9",
+    "@rango-dev/internal-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "bignumber.js": "^9.1.1",
     "rango-types": "^0.5.0"
