+++ wallets/provider-okx/package.json
@@ -32,6 +32,7 @@
     "@rango-dev/signer-solana": "^0.49.0",
     "@rango-dev/signer-sui": "^0.11.0",
     "@rango-dev/signer-tron": "^0.42.0",
+    "@rango-dev/internal-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.1-next.2",
     "@ton/core": "^0.59.0",
