+++ wallets/provider-trezor/package.json
@@ -26,6 +26,7 @@
     "@hub3js/evm": "^0.4.0",
     "@hub3js/std": "^0.3.0",
     "@rango-dev/signer-evm": "^0.45.0",
+    "@rango-dev/internal-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.1-next.2",
     "@trezor/connect-web": "^9.5.0",
