+++ wallets/provider-unisat/package.json
@@ -24,6 +24,7 @@
     "@hub3js/core": "^0.61.1",
     "@hub3js/namespaces": "^0.3.0",
     "@hub3js/std": "^0.3.0",
+    "@rango-dev/internal-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.1-next.2",
     "bitcoinjs-lib": "^6.1.7",
