+++ wallets/provider-vultisig/src/namespaces/utxo/singer.ts
@@ -1,7 +1,7 @@
 import type { SendTransactionArgs } from '../../types.js';
 import type { Transfer } from 'rango-types/mainApi';
 
-import { isZcashBlockchain } from '@rango-dev/wallets-shared';
+import { isZcashBlockchain } from '@rango-dev/internal-blockchains';
 import { type GenericSigner, SignerError, SignerErrorCode } from 'rango-types';
 
 import { vultisigZcash } from '../../utils.js';
